chore: rename suite-common/suite-sync/src/labeling params
